+++ config.h.cmake
@@ -38,6 +38,9 @@
 /* does the compiler support conditionally throwing exceptions in constexpr context? */
 #cmakedefine DUNE_SUPPORTS_CXX_THROW_IN_CONSTEXPR 1
 
+/* does the standard library provides aligned_alloc()? */
+#cmakedefine DUNE_HAVE_C_ALIGNED_ALLOC 1
+
 /* does the standard library provide <experimental/type_traits> ? */
 #cmakedefine DUNE_HAVE_HEADER_EXPERIMENTAL_TYPE_TRAITS 1