+++ dune/common/dynmatrixev.hh
@@ -3,9 +3,11 @@
 #ifndef DUNE_DYNMATRIXEIGENVALUES_HH
 #define DUNE_DYNMATRIXEIGENVALUES_HH
 
+#include <algorithm>
 #include <memory>
 
 #include "dynmatrix.hh"
+#include "fmatrixev.hh"
 
 /*!
    \file
@@ -19,32 +21,30 @@ namespace Dune {
 
   namespace DynamicMatrixHelp {
 
-    // defined in fmatrixev_ext.cpp
-    extern void eigenValuesNonsymLapackCall(
-      const char* jobvl, const char* jobvr, const long
-      int* n, double* a, const long int* lda, double* wr, double* wi, double* vl,
-      const long int* ldvl, double* vr, const long int* ldvr, double* work,
-      const long int* lwork, const long int* info);
+    using Dune::FMatrixHelp::eigenValuesNonsymLapackCall;
 
     /** \brief calculates the eigenvalues of a symmetric field matrix
         \param[in]  matrix matrix eigenvalues are calculated for
         \param[out] eigenValues FieldVector that contains eigenvalues in
                     ascending order
+        \param[out] eigenVectors (optional) list of right eigenvectors
 
         \note LAPACK::dgeev is used to calculate the eigen values
      */
     template <typename K, class C>
     static void eigenValuesNonSym(const DynamicMatrix<K>& matrix,
-                                  DynamicVector<C>& eigenValues)
+                                  DynamicVector<C>& eigenValues,
+                                  std::vector<DynamicVector<K>>* eigenVectors = nullptr
+      )
     {
       {
         const long int N = matrix.rows();
         const char jobvl = 'n';
-        const char jobvr = 'n';
+        const char jobvr = eigenVectors ? 'v' : 'n';
 
 
         // matrix to put into dgeev
-        std::unique_ptr<double[]> matrixVector = std::make_unique<double[]>(N*N);
+        auto matrixVector = std::make_unique<double[]>(N*N);
 
         // copy matrix
         int row = 0;
@@ -57,17 +57,19 @@ namespace Dune {
         }
 
         // working memory
-        std::unique_ptr<double[]> eigenR = std::make_unique<double[]>(N);
-        std::unique_ptr<double[]> eigenI = std::make_unique<double[]>(N);
-        std::unique_ptr<double[]> work = std::make_unique<double[]>(3*N);
+        auto eigenR = std::make_unique<double[]>(N);
+        auto eigenI = std::make_unique<double[]>(N);
+
+        const long int lwork = eigenVectors ? 4*N : 3*N;
+        auto work = std::make_unique<double[]>(lwork);
+        auto vr = eigenVectors ? std::make_unique<double[]>(N*N) : std::unique_ptr<double[]>{};
 
         // return value information
         long int info = 0;
-        long int lwork = 3*N;
 
         // call LAPACK routine (see fmatrixev_ext.cc)
-        eigenValuesNonsymLapackCall(&jobvl, &jobvr, &N, &matrixVector[0], &N,
-                                    &eigenR[0], &eigenI[0], 0, &N, 0, &N, &work[0],
+        eigenValuesNonsymLapackCall(&jobvl, &jobvr, &N, matrixVector.get(), &N,
+                                    eigenR.get(), eigenI.get(), nullptr, &N, vr.get(), &N, work.get(),
                                     &lwork, &info);
 
         if( info != 0 )
@@ -79,6 +81,15 @@ namespace Dune {
         eigenValues.resize(N);
         for (int i=0; i<N; ++i)
           eigenValues[i] = std::complex<double>(eigenR[i], eigenI[i]);
+
+        if (eigenVectors) {
+          eigenVectors->resize(N);
+          for (int i = 0; i < N; ++i) {
+            auto& v = (*eigenVectors)[i];
+            v.resize(N);
+            std::copy(vr.get() + N*i, vr.get() + N*(i+1), &v[0]);
+          }
+        }
       }
     }