+++ dune/common/fmatrixev.cc
@@ -65,6 +65,18 @@ extern "C" {
    **  w       (output) BASE DATA TYPE array, dimension (n)
    **          if info = 0, the eigenvalues in ascending order.
    **
+   **  work    (workspace/output) DOUBLE PRECISION array, dimension (MAX(1,LWORK))
+   **           On exit, if INFO = 0, WORK(1) returns the optimal LWORK.
+   **
+   **  lwork   (input) INTEGER
+   **          The length of the array WORK.  LWORK >= max(1,3*N-1).
+   **          For optimal efficiency, LWORK >= (NB+2)*N,
+   **          where NB is the blocksize for DSYTRD returned by ILAENV.
+   **
+   **          If LWORK = -1, then a workspace query is assumed; the routine
+   **          only calculates the optimal size of the WORK array, returns
+   **          this value as the first entry of the WORK array, and no error
+   **          message related to LWORK is issued by XERBLA.
    **
    **
    **  info    (output) long int
@@ -182,7 +194,7 @@ extern "C" {
   extern void DGEEV_FORTRAN(const char* jobvl, const char* jobvr, const long
                             int* n, double* a, const long int* lda, double* wr, double* wi, double* vl,
                             const long int* ldvl, double* vr, const long int* ldvr, double* work,
-                            const long int* lwork, const long int* info);
+                            const long int* lwork, long int* info);
 
 } // end extern C
 #endif
@@ -212,7 +224,7 @@ namespace Dune {
       const char* jobvl, const char* jobvr, const long
       int* n, double* a, const long int* lda, double* wr, double* wi, double* vl,
       const long int* ldvl, double* vr, const long int* ldvr, double* work,
-      const long int* lwork, const long int* info)
+      const long int* lwork, long int* info)
     {
 #if HAVE_LAPACK
       // call LAPACK dgeev